# docs/swagger.yaml
basePath: /api/v2
definitions:
  model.Chart:
    properties:
      chart_id:
        type: integer
      difficulty:
        allOf:
        - $ref: '#/definitions/model.Difficulty'
        example: massive
      fitting_level:
        type: number
      level:
        type: number
      level_design:
        type: string
      notes:
        type: integer
      song:
        $ref: '#/definitions/model.Song'
      song_id:
        type: integer
    type: object
  model.ChartInfo:
    properties:
      album:
        example: First Album
        type: string
      artist:
        example: Artist Name
        type: string
      b15:
        example: false
        type: boolean
      bpm:
        example: "180"
        type: string
      chart_id:
        example: 10
        type: integer
      cover:
        example: https://example.com/cover.jpg
        type: string
      difficulty:
        allOf:
        - $ref: '#/definitions/model.Difficulty'
        example: massive
      fitting_level:
        example: 13.4
        type: number
      genre:
        example: Pop
        type: string
      illustrator:
        example: Artist
        type: string
      length:
        example: "2:30"
        type: string
      level:
        example: 13.2
        type: number
      level_design:
        example: Designer
        type: string
      notes:
        example: 850
        type: integer
      song_id:
        example: 1
        type: integer
      title:
        example: Song Title
        type: string
      version:
        example: 1.0.0
        type: string
      wiki_id:
        example: w123
        type: string
    type: object
  model.ChartInfoSimple:
    properties:
      b15:
        type: boolean
      chart_id:
        type: integer
      cover:
        type: string
      difficulty:
        $ref: '#/definitions/model.Difficulty'
      fitting_level:
        type: number
      level:
        type: number
      song_id:
        type: integer
      title:
        type: string
      version:
        type: string
      wiki_id:
        type: string
    type: object
  model.Difficulty:
    enum:
    - detected
    - invaded
    - massive
    - reboot
    type: string
    x-enum-varnames:
    - DifficultyDetected
    - DifficultyInvaded
    - DifficultyMassive
    - DifficultyReboot
  model.LevelInfo:
    properties:
      difficulty:
        allOf:
        - $ref: '#/definitions/model.Difficulty'
        example: massive
      level:
        example: 14.5
        type: number
      level_design:
        example: Designer
        type: string
      notes:
        example: 1000
        type: integer
    type: object
  model.PlayRecord:
    properties:
      chart:
        $ref: '#/definitions/model.Chart'
      chart_id:
        example: 1
        type: integer
      play_record_id:
        type: integer
      rating:
        type: integer
      record_time:
        type: string
      score:
        example: 1000000
        minimum: 0
        type: integer
      username:
        type: string
    required:
    - chart_id
    type: object
  model.PlayRecordBase:
    properties:
      chart_id:
        example: 1
        type: integer
      score:
        example: 1000000
        minimum: 0
        type: integer
    required:
    - chart_id
    type: object
  model.PlayRecordInfo:
    properties:
      chart:
        $ref: '#/definitions/model.ChartInfoSimple'
      play_record_id:
        type: integer
      rating:
        type: integer
      record_time:
        type: string
      score:
        type: integer
    type: object
  model.PlayRecordResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/model.PlayRecordInfo'
        type: array
      total:
        type: integer
      username:
        type: string
    type: object
  model.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  model.Song:
    properties:
      album:
        example: First Album
        type: string
      artist:
        example: Artist Name
        type: string
      b15:
        example: false
        type: boolean
      bpm:
        example: "180"
        type: string
      charts:
        items:
          $ref: '#/definitions/model.Chart'
        type: array
      cover:
        example: https://example.com/cover.jpg
        type: string
      genre:
        example: Pop
        type: string
      illustrator:
        example: Artist
        type: string
      length:
        example: "2:30"
        type: string
      song_id:
        type: integer
      title:
        example: Song Title
        type: string
      version:
        example: 1.0.0
        type: string
      wiki_id:
        example: w123
        type: string
    type: object
...
  model.UploadFileResponse:
    properties:
      filename:
        type: string
    type: object
...
  request.BatchCreatePlayRecordRequest:
    properties:
      csv_filename:
        type: string
      is_replace:
        type: boolean
      play_records:
        items:
          $ref: '#/definitions/model.PlayRecordBase'
        type: array
      upload_token:
        type: string
    type: object
...
  request.CreateSongRequest:
    properties:
      album:
        example: First Album
        type: string
      artist:
        example: Artist Name
        type: string
      b15:
        example: false
        type: boolean
      bpm:
        example: "180"
        type: string
      cover:
        example: https://example.com/cover.jpg
        type: string
      genre:
        example: Pop
        type: string
      illustrator:
        example: Artist
        type: string
      length:
        example: "2:30"
        type: string
      levels:
        items:
          $ref: '#/definitions/model.LevelInfo'
        type: array
      title:
        example: Song Title
        type: string
      version:
        example: 1.0.0
        type: string
      wiki_id:
        example: w123
        type: string
    type: object
...
  request.UpdateSongRequest:
    properties:
      album:
        example: First Album
        type: string
      artist:
        example: Artist Name
        type: string
      b15:
        example: false
        type: boolean
      bpm:
        example: "180"
        type: string
      cover:
        example: https://example.com/cover.jpg
        type: string
      genre:
        example: Pop
        type: string
      illustrator:
        example: Artist
        type: string
      length:
        example: "2:30"
        type: string
      levels:
        items:
          $ref: '#/definitions/model.LevelInfo'
        type: array
      song_id:
        type: integer
      title:
        example: Song Title
        type: string
      version:
        example: 1.0.0
        type: string
      wiki_id:
        example: w123
        type: string
    required:
    - song_id
    type: object
...
host: api.prp.icel.site
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 'This is the API documentation for the Paradigm: Reboot Prober service.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'Paradigm: Reboot Prober API'
  version: "2"
paths:
  /records/{username}:
    get:
      description: Retrieve play records for a user based on scope (b50, best, all,
        all-charts)
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - default: b50
        description: Scope (b50, best, all, all-charts)
        in: query
        name: scope
        type: string
      - default: 0
        description: Underflow for b50
        in: query
        name: underflow
        type: integer
      - default: 50
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: 1
        description: Page index
        in: query
        name: page_index
        type: integer
      - default: rating
        description: Sort by (rating, score, record_time, etc.)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Order (desc or asce)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PlayRecordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get play records
      tags:
      - record
    post:
      consumes:
      - application/json
      description: Batch upload play records for a user
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Play records upload info
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/request.BatchCreatePlayRecordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.PlayRecord'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: Upload play records
      tags:
      - record
  /songs:
    get:
      description: Retrieve a list of all charts with their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChartInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all charts
      tags:
      - song
    post:
      consumes:
      - application/json
      description: Create a new song with its charts (Admin only)
      parameters:
      - description: Song creation info
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/request.CreateSongRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChartInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Create a new song
      tags:
      - song
    put:
      consumes:
      - application/json
      description: Update song details and its charts (Admin only)
      parameters:
      - description: Song update info
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/request.UpdateSongRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ChartInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Update an existing song
      tags:
      - song
  /songs/{song_id}:
    get:
      description: Retrieve detailed information about a single song by ID
      parameters:
      - description: Song ID
        in: path
        name: song_id
        required: true
        type: string
      - default: prp
        description: Source (prp or wiki)
        in: query
        name: src
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Song'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get single song info
      tags:
      - song
  /upload/csv:
    post:
      consumes:
      - multipart/form-data
      description: Upload a CSV file containing play records
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UploadFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Upload CSV file
      tags:
      - upload
  /upload/img:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file (Admin only)
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UploadFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Upload image file
      tags:
      - upload
...
schemes:
- https
swagger: "2.0"