# internal/controller/upload.go
package controller

import (
	"crypto/rand"
	"encoding/hex"
	"fmt"
	"net/http"
	"os"
	"paradigm-reboot-prober-go/config"
	"paradigm-reboot-prober-go/internal/model"
	"paradigm-reboot-prober-go/internal/service"
	"path/filepath"
	"strings"

	"github.com/gin-gonic/gin"
)

type UploadController struct {
	userService *service.UserService
}

func NewUploadController(userService *service.UserService) *UploadController {
	return &UploadController{userService: userService}
}

func generateRandomHex(n int) string {
	b := make([]byte, n)
	if _, err := rand.Read(b); err != nil {
		// Fallback to a less ideal but functional approach
		for i := range b {
			b[i] = byte(i ^ 0xAB)
		}
	}
	return hex.EncodeToString(b)
}

// UploadCSV godoc
// @Summary Upload CSV file
// @Description Upload a CSV file containing play records
// @Tags upload
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param file formData file true "CSV file"
// @Success 200 {object} model.UploadFileResponse
// @Failure 400 {object} model.Response
// @Failure 401 {object} model.Response
// @Router /upload/csv [post]
func (ctrl *UploadController) UploadCSV(c *gin.Context) {
	username := c.GetString("username")
	if username == "" {
		c.JSON(http.StatusUnauthorized, model.Response{Error: "authentication required"})
		return
	}

	file, err := c.FormFile("file")
	if err != nil {
		c.JSON(http.StatusBadRequest, model.Response{Error: "no file uploaded"})
		return
	}

	if !strings.EqualFold(filepath.Ext(file.Filename), ".csv") {
		c.JSON(http.StatusBadRequest, model.Response{Error: "only CSV files are allowed"})
		return
	}

	// Ensure directory exists
	if err := os.MkdirAll(config.GlobalConfig.Upload.CSVPath, 0750); err != nil {
		c.JSON(http.StatusInternalServerError, model.Response{Error: "failed to create upload directory"})
		return
	}

	filename := fmt.Sprintf("%s_b50_%s.csv", username, generateRandomHex(6))
	dst := filepath.Join(config.GlobalConfig.Upload.CSVPath, filename)

	if err := c.SaveUploadedFile(file, dst); err != nil {
		c.JSON(http.StatusInternalServerError, model.Response{Error: "failed to save file"})
		return
	}

	c.JSON(http.StatusOK, model.UploadFileResponse{Filename: filename})
}

// UploadImg godoc
// @Summary Upload image file
// @Description Upload an image file (Admin only)
// @Tags upload
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param file formData file true "Image file"
// @Success 200 {object} model.UploadFileResponse
// @Failure 400 {object} model.Response
// @Failure 401 {object} model.Response
// @Failure 403 {object} model.Response
// @Router /upload/img [post]
func (ctrl *UploadController) UploadImg(c *gin.Context) {
	username := c.GetString("username")
	user, err := ctrl.userService.GetUser(username)
	if err != nil || user == nil || !user.IsAdmin {
		c.JSON(http.StatusForbidden, model.Response{Error: "admin access required"})
		return
	}

	file, err := c.FormFile("file")
	if err != nil {
		c.JSON(http.StatusBadRequest, model.Response{Error: "no file uploaded"})
		return
	}

	// Ensure directory exists
	if err := os.MkdirAll(config.GlobalConfig.Upload.ImgPath, 0750); err != nil {
		c.JSON(http.StatusInternalServerError, model.Response{Error: "failed to create upload directory"})
		return
	}

	// Sanitize filename to prevent path traversal
	safeFilename := filepath.Base(file.Filename)
	filename := fmt.Sprintf("%s_%s", generateRandomHex(8), safeFilename)
	dst := filepath.Join(config.GlobalConfig.Upload.ImgPath, filename)

	if err := c.SaveUploadedFile(file, dst); err != nil {
		c.JSON(http.StatusInternalServerError, model.Response{Error: "failed to save file"})
		return
	}

	c.JSON(http.StatusOK, model.UploadFileResponse{Filename: filename})
}